api: Move object_storage_sstables handler to storage_manager.cc
Relocate rest_object_storage_sstables and its helpers
(require_query_param, object_storage_prefix, get_object_storage_client,
collect_object_storage_entries, list_object_storage_entries,
first_path_component) verbatim from storage_service.cc to
storage_manager.cc. The handler still reads ctx.db internally,
unchanged; only the registration site moves from
set_storage_service()'s gated(ss, rest_bind(...)) wrapper to a plain
lambda in set_storage_manager().

Drop the now-unused sstables/object_storage_client.hh and
sstables/storage.hh includes from storage_service.cc; they were only
needed by the relocated code. sstables/sstables_manager.hh stays,
it's still used by estimate_compression_ratios().

Dropping gated() here is safe: it held service::storage_service's
async gate so that storage_service::stop() would drain in-flight
requests, but this handler never touched ss - it only gained that
protection incidentally by living in the same file. It has nothing
to do with storage_service's lifecycle. The equivalent protection
for the actual dependency (sstables::storage_manager, wired up in
the next patch) would need to be a gate on storage_manager itself,
which doesn't exist; this matches the existing, already-accepted
pattern in compaction_manager.cc/tasks.cc, whose handlers likewise
capture their sharded service with no gating against its stop().
No other logic changes.
